terra-order-transaction-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
redis==4.6.0
celery==5.3.4
pika==1.3.2
requests==2.31.0
django-filter==23.3
drf-yasg==1.21.7
python-decouple==3.8
django-environ==0.10.0
pyjwt==2.8.0
django-celery-results==2.5.1
channels==4.0.0
channels-redis==4.1.0
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
pytest-mock==3.11.1
freezegun==1.2.2
factory-boy==3.3.0
coverage==7.2.7